Fixed some game logic
Still need to fix gameOver().
Request is not sent until Next Team is pressed, only then does it realise the game is actually over. A broken round will be displayed.
Could probably be solved by "onload" in the body or something, to check if the game is over between the rounds once.